Cache faqs and recent notices for topics
This page is quite slow. The FAQs and recent notice feed are extremely
cacheable, however. This makes a small difference on localhost but
should make quite a large one in prod with the high number of notices.

Requires moving some topics_controller functions to a helper, so that
we can avoid constructing @notices unless we absolutely need to write
the cache fragment. But this is also great as it factors topics#show
into a smaller thing with a unitary responsibility. Hooray!
