+++ app/views/topics/show.html.erb
@@ -7,23 +7,28 @@
   </article>
 </section>
 
-<% if @notices.any? %>
-  <section class="topic-notices">
-    <h4>Recent Notices</h4>
-    <ol>
-      <%= render @notices %>
-    </ol>
-  </section>
+<% cache "topic_#{@topic.id}_notices" do %>
+  <% notices = construct_recent_notices %>
+  <% if notices.any? %>
+    <section class="topic-notices">
+      <h4>Recent Notices</h4>
+      <ol>
+        <%= render notices %>
+      </ol>
+    </section>
+  <% end %>
 <% end %>
 
-<section class="topic-faqs" id="topic-faqs">
-  <h3 class="title">Frequently Asked Questions</h3>
-  <div class="wrapper">
-  <% @topic.relevant_questions.each do |question| %>
-    <article>
-      <div class="question"><%= question.question %></div>
-      <div class="answer"><%=raw question.answer_html %></div>
-    </article>
-  <% end %>
-  </div>
-</section>
+<% cache "topic_#{@topic.id}_faqs" do %>
+  <section class="topic-faqs" id="topic-faqs">
+    <h3 class="title">Frequently Asked Questions</h3>
+    <div class="wrapper">
+    <% @topic.relevant_questions.each do |question| %>
+      <article>
+        <div class="question"><%= question.question %></div>
+        <div class="answer"><%=raw question.answer_html %></div>
+      </article>
+    <% end %>
+    </div>
+  </section>
+<% end %>
